data_guardian/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-guardian"
version = "0.1.0"
description = "Data Guardian: AES-256-GCM + RSA-OAEP + Ed25519 (pure Python backend/OOP)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "cryptography>=42.0.0",
  "typer>=0.12.0",
  "tqdm>=4.66.0",
  "PyJWT>=2.8.0",
  "prometheus-client>=0.20.0",
]

[project.scripts]
data-guardian = "data_guardian.cli:app"
